Avance 4

+++ pag_inicio/src/main/java/pe/edu/uni/pag_inicio/service/IUsuariosService.java
@@ -15,5 +15,5 @@ public interface IUsuariosService {
 
     List<Usuarios> getAllUsers();
 
-    void deleteUser(int userId);
+    void deleteUser(Long userId);
 }
